Refactor resize out of project page
